pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantengine"
version = "0.2.0"
description = "QuantEngine: GPU-first vectorized backtesting engine"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "QuantEngine Contributors" }]
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.0.0",
  "polars>=0.20.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0",
  "rich>=13.7.0",
  "click>=8.1.0",
]

[project.optional-dependencies]
engine = [
  "cupy-cuda12x>=13.0",
  "cudf-cu12>=24.0",
  "optuna>=3.6",
  "nevergrad>=1.0",
  "deap>=1.4",
  "plotly>=5.18",
  "matplotlib>=3.8",
  "torch>=2.2",
  "numba>=0.59",
]

[project.scripts]
quantengine = "quantengine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies.dev]
ruff = ">=0.8.0"
pytest = ">=8.0"
pytest-cov = ">=5.0"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "SIM",  # flake8-simplify
  "T20",  # flake8-print (disallow print in src/)
]
ignore = [
  "E501",   # line length handled by formatter
  "SIM108", # ternary operator preference
  "SIM102", # nested-if in Numba @njit functions is intentional for compatibility
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["T20"]   # print allowed in scripts
"tests/*" = ["T20"]     # print allowed in tests

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-q", "-m", "not slow", "--tb=short"]
markers = [
  "slow: 真实数据验收测试",
]
